[main] Update GitHub Actions workflow for Docker image publishing
This commit modifies the existing GitHub Actions workflow (docker_build_test.yml) to include the necessary steps for publishing the Docker image to the GitHub Container Registry. The workflow now authenticates with the registry using the provided secret (githubdockerregistry) and pushes the tagged Docker image to the repository's container registry.

These changes enable the seamless publishing of the Docker image to the GitHub Container Registry, facilitating easier access and utilization of the image by other users or workflows.

add missing requirements.txt