Fix Docker image tagging and pushing error due to uppercase repository name
This commit addresses an error that occurred during the Docker image tagging and pushing process in the `.github/workflows/docker_build_test.yml` workflow. The error was caused by using an uppercase repository name in the reference for tagging and pushing the image to the GitHub Container Registry.

The specific error message encountered was:
"Error parsing reference: 'ghcr.io/Aricg/dns-text-record-monitor:latest' is not a valid repository/tag: invalid reference format: repository name must be lowercase."

To resolve this issue, the code has been updated to convert the repository name to lowercase using the `tr` command. The Docker image is now correctly tagged and pushed to the GitHub Container Registry using the lowercase repository name.

By making this change, the workflow can successfully tag and push the Docker image without encountering the "invalid reference format" error.